constellation_12_class/requirements.txt
tensorflow==2.10.0
pillow==9.0.1
